pyproject.toml
[project]
name = "medimind-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu==1.12.0",
    "llama-index==0.14.0",
    "llama-index-embeddings-zhipuai==0.3.1",
    "llama-index-llms-zhipuai==0.4.1",
    "llama-index-readers-file==0.5.4",
    "llama-index-vector-stores-faiss==0.5.1",
    "numpy==2.2.6",
    "openpyxl==3.1.5",
    "pandas<2.3.0",
    "pydantic==2.12.3",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "sentence-transformers==5.1.1",
    "tiktoken==0.12.0",
    "torch==2.9.0",
    "transformers==4.57.1",
    "xlrd>=2.0.0",
]
